scanner mit umlauten gefixt, cover in function, volume wird in cookie gespeichert, scan mit option  datenbank löschen

+++ folder_v2.php
@@ -1,4 +1,5 @@
-<html>
+﻿<html>
+<meta http-equiv="content-type" content="text/html; charset=UTF-8">
 <head>
 	<link rel="stylesheet" href="./themes/base/jquery.ui.all.css">
 	<script src="./js/jquery-1.9.1.js"></script>
@@ -52,20 +53,21 @@ require_once('config.inc.php');
 require_once('./getid3/getid3.php');
 mysql_connect(DBHOST, DBUSER,DBPASS) OR DIE ("NICHT Erlaubt");
 mysql_select_db(DBDATABASE) or die ("Die Datenbank existiert nicht.");
+mysql_query("SET NAMES 'utf8'");
 
 mysql_query("TRUNCATE `scanner_log`");
-/*
+if($_REQUEST['truncate']=="yes") {
 mysql_query("TRUNCATE `album`");
 mysql_query("TRUNCATE `artist`");
 mysql_query("TRUNCATE `title`");
-*/
+}
 mysql_query("INSERT INTO scanner_log (starttime) VALUES (NOW())");	
 $dirs = array();
 // Initialize getID3 engine
 $getID3 = new getID3;
 
 //$DirectoryToScan = MUSICDIR; // change to whatever directory you want to scan
-$DirectoryToScan = $_REQUEST['scandir'];
+$DirectoryToScan = utf8_encode($_REQUEST['scandir']);
 
 $out  = '';  
 function scan($folder){  
@@ -91,6 +93,7 @@ function scan($folder){
 echo scan($DirectoryToScan);
 $dirs = (explode(';', $out));
 $dirstoscan = count($dirs);
+echo "Durchsuche $$DirectoryToScan ...<br>";
 echo $dirstoscan -1 ." Ordner gefunden. Starte Suche...<br>";
 /*
 for ($i = 0; $i <= $dirstoscan; $i++) {
